1.19.4

+++ src/main/java/trackapi/lib/Util.java
@@ -9,7 +9,7 @@ import trackapi.compat.MinecraftRail;
 
 public class Util {
 	private static ITrack getInternalTileEntity(final Level world, Vec3 pos, boolean acceptMinecraftRails) {
-		final BlockPos bp = new BlockPos(Math.floor(pos.x), Math.floor(pos.y), Math.floor(pos.z));
+		final BlockPos bp = new BlockPos((int)Math.floor(pos.x), (int)Math.floor(pos.y), (int)Math.floor(pos.z));
 		BlockState bs = world.getBlockState(bp);
 		
 		if (bs.getBlock() instanceof ITrackBlock) {
